+++ comet/utils/log_levels.py
@@ -32,6 +32,12 @@ CUSTOM_LOG_LEVELS = {
         "loguru_color": "<fg #71d6d6>",
         "no": 30,
     },
+    "SESSION": {
+        "color": "#71d6d6",
+        "icon": "🔒",
+        "loguru_color": "<fg #71d6d6>",
+        "no": 30,
+    },
     "BACKGROUND_SCRAPER": {
         "color": "#5fba64",
         "icon": "🏭",